mcp: carve tools_display.rs (pure move)
The workspace/display tool implementations move into their own impl
block: browser workspaces, display take/release + user-display grants,
shared-view emit/show/hide/focus/capture, computer-use actions,
screenshots/screen reading, frames, and live-audio spawn — with 10
tests. The emptied source impl shell is dropped.

Usual glue; purity-checked; mcp:: tests 178/178.
